+++ core/auth/remote/storage-remote.ts
@@ -4,17 +4,8 @@
  * For web clients that need to share state with Node apps.
  */
 
-import type { OAuthStorage } from "../storage.js";
-import type {
-  OAuthClientInformation,
-  OAuthTokens,
-  OAuthMetadata,
-} from "@modelcontextprotocol/sdk/shared/auth.js";
-import {
-  OAuthClientInformationSchema,
-  OAuthTokensSchema,
-} from "@modelcontextprotocol/sdk/shared/auth.js";
-import { createOAuthStore, type ServerOAuthState } from "../store.js";
+import { OAuthStorageBase } from "../oauth-storage.js";
+import { createOAuthStore } from "../store.js";
 import { createRemoteStorageAdapter } from "../../storage/adapters/remote-storage.js";
 
 export interface RemoteOAuthStorageOptions {
@@ -33,135 +24,17 @@ export interface RemoteOAuthStorageOptions {
  * Stores OAuth state via HTTP API (GET/POST/DELETE /api/storage/:storeId).
  * For web clients that need to share state with Node apps (TUI, CLI).
  */
-export class RemoteOAuthStorage implements OAuthStorage {
-  private store: ReturnType<typeof createOAuthStore>;
-
+export class RemoteOAuthStorage extends OAuthStorageBase {
   constructor(options: RemoteOAuthStorageOptions) {
-    const storage = createRemoteStorageAdapter({
-      baseUrl: options.baseUrl,
-      storeId: options.storeId ?? "oauth",
-      authToken: options.authToken,
-      fetchFn: options.fetchFn,
-    });
-    this.store = createOAuthStore(storage);
-  }
-
-  async getClientInformation(
-    serverUrl: string,
-    isPreregistered?: boolean,
-  ): Promise<OAuthClientInformation | undefined> {
-    const state = this.store.getState().getServerState(serverUrl);
-    const clientInfo = isPreregistered
-      ? state.preregisteredClientInformation
-      : state.clientInformation;
-
-    if (!clientInfo) {
-      return undefined;
-    }
-
-    return await OAuthClientInformationSchema.parseAsync(clientInfo);
-  }
-
-  async saveClientInformation(
-    serverUrl: string,
-    clientInformation: OAuthClientInformation,
-  ): Promise<void> {
-    this.store.getState().setServerState(serverUrl, {
-      clientInformation,
-    });
-  }
-
-  async savePreregisteredClientInformation(
-    serverUrl: string,
-    clientInformation: OAuthClientInformation,
-  ): Promise<void> {
-    this.store.getState().setServerState(serverUrl, {
-      preregisteredClientInformation: clientInformation,
-    });
-  }
-
-  clearClientInformation(serverUrl: string, isPreregistered?: boolean): void {
-    this.store.getState().getServerState(serverUrl);
-    const updates: Partial<ServerOAuthState> = {};
-
-    if (isPreregistered) {
-      updates.preregisteredClientInformation = undefined;
-    } else {
-      updates.clientInformation = undefined;
-    }
-
-    this.store.getState().setServerState(serverUrl, updates);
-  }
-
-  async getTokens(serverUrl: string): Promise<OAuthTokens | undefined> {
-    const state = this.store.getState().getServerState(serverUrl);
-    if (!state.tokens) {
-      return undefined;
-    }
-
-    return await OAuthTokensSchema.parseAsync(state.tokens);
-  }
-
-  async saveTokens(serverUrl: string, tokens: OAuthTokens): Promise<void> {
-    this.store.getState().setServerState(serverUrl, { tokens });
-  }
-
-  clearTokens(serverUrl: string): void {
-    this.store.getState().setServerState(serverUrl, { tokens: undefined });
-  }
-
-  getCodeVerifier(serverUrl: string): string | undefined {
-    const state = this.store.getState().getServerState(serverUrl);
-    return state.codeVerifier;
-  }
-
-  async saveCodeVerifier(
-    serverUrl: string,
-    codeVerifier: string,
-  ): Promise<void> {
-    this.store.getState().setServerState(serverUrl, { codeVerifier });
-  }
-
-  clearCodeVerifier(serverUrl: string): void {
-    this.store
-      .getState()
-      .setServerState(serverUrl, { codeVerifier: undefined });
-  }
-
-  getScope(serverUrl: string): string | undefined {
-    const state = this.store.getState().getServerState(serverUrl);
-    return state.scope;
-  }
-
-  async saveScope(serverUrl: string, scope: string | undefined): Promise<void> {
-    this.store.getState().setServerState(serverUrl, { scope });
-  }
-
-  clearScope(serverUrl: string): void {
-    this.store.getState().setServerState(serverUrl, { scope: undefined });
-  }
-
-  getServerMetadata(serverUrl: string): OAuthMetadata | null {
-    const state = this.store.getState().getServerState(serverUrl);
-    return state.serverMetadata || null;
-  }
-
-  async saveServerMetadata(
-    serverUrl: string,
-    metadata: OAuthMetadata,
-  ): Promise<void> {
-    this.store
-      .getState()
-      .setServerState(serverUrl, { serverMetadata: metadata });
-  }
-
-  clearServerMetadata(serverUrl: string): void {
-    this.store
-      .getState()
-      .setServerState(serverUrl, { serverMetadata: undefined });
-  }
-
-  clear(serverUrl: string): void {
-    this.store.getState().clearServerState(serverUrl);
+    super(
+      createOAuthStore(
+        createRemoteStorageAdapter({
+          baseUrl: options.baseUrl,
+          storeId: options.storeId ?? "oauth",
+          authToken: options.authToken,
+          fetchFn: options.fetchFn,
+        }),
+      ),
+    );
   }
 }
